+++ RentTeslaServer/RentTeslaServer/Domain Layer/ModelDtos/CarDto.cs
@@ -3,7 +3,7 @@
     public class CarDto
     {
         public int Id { get; set; }
-        public double DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
         public string Name { get; set; }
         public string Motor { get; set; }
         public int Range { get; set; }
