+++ RentTeslaServer/RentTeslaServer/Domain Layer/ModelDtos/CarManagmentDetailDto.cs
@@ -10,7 +10,7 @@ namespace RentTeslaServer.Domain_Layer.ModelDtos
         public int Id { get; set; }
         public bool IsPrepared { get; set; }
         public bool IsFree { get; set; }
-        public double DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
         public string Plates { get; set; }
         public CarTypeDto CarTypeDto { get; set; }
         public CarRentalDto CarRentalDto { get; set; }
